refactor: move InterfaceT into vyper/semantics/types/module.py
it makes more sense since it is fairly closely coupled with ModuleT.
