+++ vyper/semantics/types/__init__.py
@@ -2,9 +2,10 @@ from . import primitives, subscriptable, user
 from .base import TYPE_T, KwargSettings, VyperType, is_type_t
 from .bytestrings import BytesT, StringT, _BytestringT
 from .function import MemberFunctionT
+from .module import InterfaceT
 from .primitives import AddressT, BoolT, BytesM_T, DecimalT, IntegerT
 from .subscriptable import DArrayT, HashMapT, SArrayT, TupleT
-from .user import EnumT, EventT, InterfaceT, StructT
+from .user import EnumT, EventT, StructT
 
 
 def _get_primitive_types():
